pyproject.toml
[tool.poetry]
name = "python-fastapi-manager"
packages = [{include="fastapi_manager"}]
version = "0.1.0"
description = ""
authors = ["AstralMortem <chaliukvladyslav@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
dynaconf = "^3.2.6"
tortoise-orm = "^0.21.6"
typer = "^0.12.5"
uvicorn = "^0.30.6"
aerich = "^0.7.2"
asyncpg = "^0.29.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
